feature/eugen/added-tests-to-check-toolbar-items * Added toolbar scenario for newly added elements; * Added FormToolbar default one; * Fixed `FormToolbar` location.
