+++ web-interactions/src/main/java/com/brizy/io/web/interactions/page/editor/container/components/type/form/field/factory/FormField.java
@@ -5,7 +5,6 @@ public interface FormField<T> {
     void setToolbarProperties(T properties);
 
     default void configureFormField(T properties) {
-
     }
 
     void duplicate();
