+++ web-interactions/src/main/java/com/brizy/io/web/interactions/page/editor/container/components/type/form/type/Email.java
@@ -4,10 +4,10 @@ import com.brizy.io.web.common.dto.element.properties.form.FormProperties;
 import com.brizy.io.web.common.dto.element.properties.form.field.type.EmailFieldProperty;
 import com.brizy.io.web.interactions.locators.editor.workspace.section.container.item.ItemLocators;
 import com.brizy.io.web.interactions.locators.editor.workspace.section.container.item.toolbar.ToolbarLocators;
-import com.brizy.io.web.interactions.page.editor.container.components.toolbar.variations.FormToolbar;
 import com.brizy.io.web.interactions.page.editor.container.components.toolbar.variations.form.type.AbstractField;
 import com.brizy.io.web.interactions.page.editor.container.components.toolbar.variations.form.type.EmailField;
 import com.brizy.io.web.interactions.page.editor.container.components.type.form.field.factory.FormField;
+import com.brizy.io.web.interactions.page.editor.container.components.type.form.toolbar.FormToolbar;
 import com.microsoft.playwright.Frame;
 import com.microsoft.playwright.Locator;
 import lombok.AccessLevel;
